Program now detects overlapping while resizing objects
Added global variables that saves the original properties of an entity before resizing while in "mdown()" and then lets the user resize freely when the program is in "mmoving()". and then when the resizing is done, the "mup()" reverts the entity to the original state if overlapping was detected and sends an error-message.