a newly created course version is displayed in "copy content from" dropdown when trying to copy from an existing course version
before it wasn't possible that when a new course version was created it wasn't copied over or not showing in the drop down but now it works. By adding another else if for CPYVRS and correcting a syntax error it now works. Also, the new function copyCourseVersion is being tested out.