Fixed timer for github repo fetch
The cooldown was displayed after simply clicking the github repo button in the navbar. This has been changed to only trigger after the save button in the popup is clicked. If the input field is left empty, the cooldown will not trigger.
There doesn't seem to be any actual validation for the input field though. As long as you enter something, the functions will trigger. Having an actual validation that not only checks the format, but also that it contains an actual working repo link would probably be outside the scope here.

Solution:
The function that that starts and displays the timer is "resetGitFetchTimer()". This function call was moved from navheader.php to sectioned.js. To be able to use the php-variable it needs to be initiated as a js-variable before sectioned.js is loaded. Until another course is visited the variable will remain static, which is fine. Since changing uid would mean relogging and then loading a new course, which means navheader.php would be loaded again and the value would therefore be initiated with the new value.

For testing:
Go to a course, click the repo button. Click save without entering anything in the field. Nothing happening? Success!
Now enter something in the field, click save. Now hover over the repo-button. The cooldown should now be displayed.