Added try/catch blocks in courseedservice.php
Added try/catch blocks in courseedservice.php to accommodate to both legacy data and the newer database. For instance, with the try/catch blocks, different queries will run depending on if there is a githubURL or not. This was implemented because the old database did not contain the column for githubURL, whereas the new database had it.