Removed nesting
I removed the nested checks and added "NONE!" as the default value of debug. I also made it so logging is performed after the file and database has been updated, so that no log is produced if something were to fail. Add the current version of retrieveUsername_ms.php to test and replace the "echo json_encode($debug)" with "echo $debug".