CSS styling added and changed
CSS styling added and changed, such as indentation, removal of inline styling and adding classes for styling. Currently, the color scheme is adjusted to both themes but the additional changes to be made surrounds the difference in margin of text and cells of the table which are offset differently in the light and dark mode.