Updated margin and style for forms on 3 files
There was a conflict with the `Add user` form in accessed and all other forms in courseed- and sectioned-, duggaed.php. The solution was to accept and implement the style without the extra 10px margin.
Extra: Added missing style class to submit button and missing body element in `Github Moment` form in sectioned.php which was mor apparent without the 10px. Changed flexwrapper to inputwrapper in `Edit Dugga` to ensure it follows the style of the other forms.