Updated thrashcan and deletion functionality
When the user clicks the trashcan item, the "confirmationBox" opens up. I've implemented so the clicked dugga get pushed to the "selectedItemList" when the user clicks "Yes" in the confirmation window. For safety reasons have I included an if-statement that don't push the dugga-id if it's already inside the "selectedItemList" array.

When yes is clicked, the function "deleteItem" deletes all the items in the array "selectedItemList".

I have left some commented code that could be helpful for testers. If everything work properly the commented code should be removed.

Commented code is on line: 650, 651, 653, 671 and 672