Changed css for the loginbox
Changed the max width to fit content in the loginbox and set the logintable to have some margin to look a bit nicer.
The only thing that is now is that on some places where the loginbox-class is used as css when the content inside is not filled wider than 200px (min-width) it get pretty small. Attached some pictures where it is changed with the new css.