Applied new css and functions
Added a function that removes classes "even" and "odd" from the table-rows that are rendered and adds them to the new ones that are rendered. Also puts the class "collapsed" to the table-rows that are hidden (collapsed) so they are not affected by the count of even odd as it was before. Had to make the function as an onclick property on the table-headings so it runs every time it sorts the table and in the function "returnedAccess" so it applies from the start of the rendering as well.
Added correct css-variables to the table background so it works in darkmode as well and therefor removed the adding of class "changeColorInDarkModeTable" because this is now unnecessary .