Generalized validation texts
Edited the css rules that apply to the formdialog and formdialogtext classes. In the previous weekly branch I had to remove some specific rules to specific classes, these appear to be removed in this branch.

The new "dialogwrapper" is intended to contain the inputfields and the formdialog div. Now, when doing this to various inputs, the validation texts appear where they are intended to. I know of a few special cases that needs to be considered and I'm currently working on those.