Added changes from old branch to new branch
Added changes from old branch to new branch due to problems with the weekly merge, which had resulted in reverting a few commits back (including the one for the issue #15473). Therefore a new branch had to be created which included the adjustments to which fixed the problem of the weird positioning of the dropdown for the new item "+"- button.