Deleted getdeletedlistentries_ms.php (readRemovedListentries_ms.php)
This microservice does makes a query to the database to retrieve listentries with visible = 3.
It never uses the result in the current state. Furthermore, retrieveSectionedService already returns all listentires, which includes those mentioned. To retrieve all listentries the microservice getListEntries may be used instead. This makes the getDeletedListEntries microservice redundant.