Applied css and function for changing color
Created a function that looks at the rows that are shown every time it re-renders and apply the even and odd classes to them. Removed the class "changeColorInDarkModeTable" that is no longer needed because of the use of correct variable in the style.css and applies it in darkmode as well.