Refactoring done on all files which include "loginBox" and "loginBoxheader"
Refactoring done on all files which include "loginBox" and "loginBoxheader". This was already done with a previous pull request, but the weekly merge broke certain elements and parts of the application. Due to this, I had to fetch the most recent version of the weekly branch (with all the new changes that unfortunately regressed after the weekly merge). Therefore, The refactoring had to be done manually again.