Renamed getGroupValues_ms.php -> readGroupValues_ms.php
Renamed the microservice and everywhere it is called in the system. Also renamed the test for the microservice, along with the function name and the url.

Note, I´m not updating the names for the microservice in the database_related_micro_services since it will be obsolet when the documentationen is done.

Also updated documentation for the ms tests.