Updated blacktheme.css
There was a margin on the select element, that has been removed. There was also a background color not using the predefined color variables. I guess it's used for something, but it was applied to the formBox select as well, so I removed that. It seems to be working fine.

I suspected the loginbox might have had consequences, but it didn't. It looks like it dide before the changes.