fixed tests
There were some error messages from sending arrays in $data with curl. So I deleted the array and only sent the variables and it seemed to work.

For the errors with the failing tests, the the formatting of the output had changed but the contents were the same. So all I did was update the expected output.