Fixed option enabling
Made it possible to select data from the UI that gets stored to a json string that sends the data to installer_engine.php. This works as expected at the moment, but fails at one step found by the backend, this will be fixed in another issue for the backend.