Removed logging and changed array initialization for #16167
This line "lid = item_lid ? item_lid : $("#lid").val();" was probably intended for a loop iterating through elements and changing their class. That function is illogical and not working at all. What it did achieve though was to initialize the array "lid" which is then used for further functionality. I moved the initialization outside to make it more clear. The other styling and the loop will be commented out and should be fixed in another issue.