Basic mousetrail added in canvasanimation.js (#16172)
I've made the basic functions of a trail for the pointer. This servers as a good foundation as the canvas works and the mouse trace and animation works.

I intend to make the dots into something more interesting and education-related. I also want to fine tune some fade outs and so on.