Created array for order of divs (#16179)
In short. I started by adding a margin to the top of each header, when I realized that every header except the first one should have a margin, so it is necessary to separate them. Since every row is a table within a div (not a TR in a single table) this couldn't be done with CSS alone.

The code creates two arrays, for now. Both will be a list of divs that are direct descendants of "sectionlistc" - which gives an ordered list where each item represents a row in the visible list.

One includes the entire ID of each div, and one includes a shorter version like "header" and "moment".

This list will be the base to apply styling depending on type of element and order of occurence.