Added conditional margin to header (#16179)
Removed the array of short ID's as it didn't help me, which I thought it would. Therefore removed the code that used regex and such to create this array.

Added a loop that gives every header except for the first one a margin. This will make every heading and its "subitems" look grouped up.