Finished logic of styling for Sectionlist (#16179)
Added logic to the styling and grouping, along with necessary class and CSS. This is based on the assumptions made from the prod version of canvas. Essentially, a header seems optional and a section seems to be used within a moment.

The added logic does the following:
1. Every header div except the first one gets a top margin
2. Every row div gets a left and a right border
3. Every moment div that isn't preceded by a header gets a small top margin
4. Every row div that precedes a moment div but isn't a header gets a bottom border
5. Every row div that precedes a header row also gets a bottom border
6. The very last row div in the array gets a bottom border
7. Every row div except for the header gets a class for shorter width

This will give us a box-like interface where a header is considered a group, and a moment (that includes sections) is considered a subgroup. If however it turns out that a section should be a parent of a moment, then this could easily be flipped.