Functionality to hide elements added to sectioned.js (#16181)
For now, this is a simple function that hides and shows a specific element in the list. It took some time to get here, but this will be a good starting point to start implementing the full functionality.

I'll add a comment with more useful information on how this works.