Togglebutton and functionality added to sectioned.js (#16181)
The hidebutton and the showbutton have been removed by using display:none. This might be temporary. If commenting them out the page won't load properly, so I'll check into that later.
They have been replaced by a button for toggling.

The button has an eventlistener targeting it and this eventlistener will toggle a variable called showHidden. This will also trigger a function called toggleHidden which manipulates the classes on each element in order to show/hide them.

The biggest issue here was that we want to hide the div-container for each row, but the hidden class is only used on child-elements of the div, so we have to identify those elements and then target their closest ancestor div.

toggleHidden also runs everytime the function "returnedSection" runs as this seems to run each time an item is dragged and moved, which caused the hidden elements to reappear whenever an item was moved, even if showHidden was set to false.

The problem now is that if an item is added on the page, the hiding/showing doesnt seem to work. I do have an idea of what it might be though, next commit.