Toggling now works after editing items (#16181)
When editing an item, the toggler stopped working. This was due to the eventhandler being defined inside "returnedSection()" which is called each time an item is edited. Each time this was called, a new instance of the eventhandler was created, which caused it to run twice. This made the toggler run twice and therefore cancel out. If editing an additional item, it ran thrice, which made it work again, and so on.

By moving its definition outside (and before "returnedSection") and creating it by referring to its function with "addEventListener('click', toggleButtonClickHandler);" this ensured that only one instance was ever created.

One might think it could simply have been defined and created outside the "returnedSection()" but that doesn't seem possible, because when returnedSection ends, the elements can no loner be targeted. That's why this has to be initiated within the function.

Anyway. The toggler should now work as expected in all cases.