Added button that can toggle between three modes (#16185)
The plan is to toggle between three modes if possible (rather than just two). I've added a button with a logo. Also added an eventlistener inside returnedSection that calls for a function that handles the toggling. As for now, it only changes the title attribute of the button, but it seems like a good base to build upon for changing the swimlane view.