Edited toggle function and added clarity to it (#16185)
When toggling now, the console accurately states which mode is activated, and the button accurately tells which mode you can toggle to. In the php-file the title is set to "Toggle to scroll" since the page always loads in "normal" mode, and the next option in the togglefunction is the scroll view.

All that's missing now is the appropiate styling for each case.