Added conditions to togglebutton (#16185)
A calculation is now done to decide wether the button should be shown or not. Toggling only meaning if the swimlane diagram is too large. Unless it has more height than 70vh, the button will not be shown.

Furthermore, since dragging elements in the list makes returnedSection run, we needed a copy of the function in the event handler to make sure that the chosen view doesn't reset. By adding the switch at the bottom of returnedSection, as a slighlt alternate version of the event handler, we can initialize the view depending on what the user has already chosen, since it's a global variable.