Original bug seems to be fixed
Original bug seems to be fixed. Unsure why this was in to do. While working on this i found that all popup windows (Edit course create, new course in courseed aswell as 20 different popup windows) have a class (display_none) that for some reason was ignored. The current solution for that is that i put back  style="display:none;" so i could test the original problem