Fixed so that the installer dosent give me any error
Fixed so that the installer dosent stop and show error when it shouldnt. Had to change the name of one file back to the original name that was in the old installer.

In the process of doing this i had to make a install log to see what was going on. Will fix so the user can download it before submiting (this will fix issue #16153)