Removed IERelation restriction from updatepos()
Removed the condition that prevented IE Relations from having resize nodes. IERelation element is now able to be resized.

The resizing doesn't work properly and needs to be adjusted before issue is solved. Added comment in diagram.js as a reminder to where the solution probably needs to be implemented.