Added "helper" for calculating text-length and adjusting ER-element size
Added a function in the helper "element.js" to calculate the length of the input text. This function can later be reused for other elements.
Also updated the drawElementEREntity-function in "element.js" in the draw-folder. The ER-entity now adjusts to the length of text the user inputs, and the weak entity border also adjusts to the new size.