Updated daywidth
Updated daywidth so now the collums fill the entire svg. Due to the hardcoded nature of the weekwidth calculation there is a possibility that tihis solution only works when theres 9 weeks in a course. I havn't found another course with the same layout where my theory can be tested unfortunately