Changed IERelation defaults, fixed resizing issues
Changed the IERelation defaults, making height half of width, and minHeight half of minWidth.

Made only corner nodes available for the element, since it's based on a circle, and other circle elements only allow for corner resizing.

Implemented special resizing for IERelation, making the height increase half as much as the width; alongside including the element in the equal resizing.

Refined the way the shape was drawn, making the shape coherent when resized.