daywidth dynamic to support any number of weeks in swimlane view #16358
Replaced the previously hardcoded daywidth (12.7) with a dynamic calculation based on the container width and number of weeks in the course. This allows the swimlane graph to fill the entire width regardless of the course length.

The solution should work for varying number of weeks, but it's difficult for me to test on my mac don't know why