Had to change the database temporary
I had to change the database since there is no documentation for the microservices yet. This means that I can't use id as a foreign key for now. So at the moment I will only insert the name of the microservice, instead of joining it with the microservice table to get the id.