Consistent information for keybinds
All element/entities' keybinds are now updated to be consistent. However, the previous solution where every entity had their own tooltipp-name (such as "tooltip-PLACE_ATTRIBUTE") was not working as they seem to not exist, for this reason all the keybinds that are used for the first element/entity is used for all in that row ("tooltip-PLACE_ATTRIBUTE" is swapped to "tooltip-PLACE_RELATION"), even though not all of them adhere to the "relation"-type - this is to make all keybinds appear and remain consistent for the user.