Adjusted ER line rendering
The cardinalities were being hidden in ER relation types, this is now fixed.

Removed the separate logic recursiveERRelation() and replaced it with getLineAttributes() to streamline how lines are calculated for all diagram types and refactored getLineAttributes() to return clear coordinates and offsets.

After the cardinality issue was resolved there was a gap that was occuring between the ER relation entity and the line so a small offset tweak was also made to avoid this.