Created setupEndpointDirectory.php that removes, installs and fills the database
Was first going to do this in installEndpointDb.php but I decided to create a separate file instead. This means that you still can reinstall the database manually if needed. I also realized that there was a typo in the fillDependenciesDb.php file title that has now been changed (missed an n).