Updated and added functionality to save buttons
Added functionality to the "save as" button, now functioning like the old save button did.

Updated functionality for the "save" button, should now autosave without prompting to name the file.

There are still minor things that need to be added, as well as bugs and such that need to be fixed before doing a pull request.