Removed hashtag symbol
Removed hashtag symbol for correct syntax between L2184-2573 based on PR #16894. Highscore pop-up dialog broke because of wrong syntax and is now fixed.

getElementById() does not need "#" symbol to retrieve id of element. If have "#" then it either ignores the # and tries to retrieve the element by id or code breaks.