Fixed issues with quicksave function and improved keybind generation
Fixed the previous issues with the quicksave function. Keybindings now work as intended and call upon the correct functions.

Updated the keybind tooltip generation to include SHIFT and ALT.

Save now instantly saves to the last file saved to. If there was no last file saved to, it prompts the user to make one, working like if you had pressed save as.

Save as now lets the user choose the name for the file they want to save to, as well as checking for overwrites and alerting the user about it.