Updated diagram.css
Added a media query that checks wether to apply those specific css-styles when the width of window is 414px or less. The styles for the icon-wrapper is important because it sort of toggles the icon-wrapper displaying or not based on the width of the window. The other styles listed below icon-wrapper in the media query is important because it eliminates most of the causing overflow that were before.