Removed quicksave call in keydown eventListener and added relevent comments
Removed the call to the function quickSaveDiagram in the keydown eventListener since it triggered the entire time the button was held down. It now functions like centerCamera and only triggers the function on key release.

Also added relevent comments to explain functionality and point out that functions that are only supposed to trigger once per press need to be called upon in the keyup eventListener