Updated diagram.js and toggle.js
I have now fixed the rulers position so it adapts to both desktop and mobile version. So, when you toggle on both version they are positioned correctly based on their environment (other elements e.g. navbar on desktop). And if you resize the window they are positioned correctly so you dont need to toggle the ruler to make the position correct.

PS: The code is now auto formatted/indented