Renamed hideErrorCheck and removed duplicate keylistener
Renamed the function hideErrorCheck() to showErrorCheck(), since the current name was a bit misleading for what it was actually doing.

Also removed the keylistener in the function which added a duplicate keybind registration in a very situational case when using diagram.php from the showDugga.php page and not being logged in to lenaSYS.

Added extra documentation to the showErrorCheck function to explain what it does. Although it doesn't seem to function properly right now.