Updated conditions for useDelta = False to prevent odd behaviour
When the element snaps back after having gone out of the viewport it still counts as a selected element. The functions for moving the element and updating its visual position are different. Since the element is still selected after snaps back the program behaved as if it was still being moved when the mouse was held down. supposed to be moved relative to the cursor position. Updating the if statement by adding extra conditions where useDelta is false fixes this.