Added data-attributes and created a array of tooltip ids
I have added necessary data-attributes for every element that has a corresponding tooltip, they are used for mapping the correct element with its tooltip information. I have also created a array of tooltip ID:s that are going to be used when writing the keybind in text.

Have also commented the code