Updated element height handling in centering of camera
The way the centerCamera() function handled getting the largest y-value before didn't account for the fact that the height of certain element types wasn't stored in the element itself. This caused issues when trying to center the camera on certain example diagrams.

This has now been fixed by taking the heights of those elements from the special arrays created for them.