Updated changeState()-function
The changeState()-function has been fixed so that it can handle a name-change or color-change before changing the "variable" (state) of an element (previously not possible to change these and then change the state). Now it handles any change to the selected elements state, and then redraws the canvas to display the change, making it possible to change name or color and then change the state.