Implemented fab-button for optionsmenu
The button is there and it works. What's left to fix is the close-button and to fix so it hides when the user is on desktop. As it is now, the button is visible all the time.

I also changed the positioning a little on the diagram-fab-button so they would match. The diagram-fab-button was a little much to the left when the yellow optionmenu is not visible anymore.