Removed the mobile view of the rectangular fab button
Removed the mobile view of the rectangular fab button.

The round fab button for mobile was also removed, as it interfered with the rectangular version: the text of each item did not display correctly.

Note: It was decided that the round fab button will be removed entirely from the mobile version which is handled in a separate issue (info from group leader). Additionally, not all items were visible when using the round fab button.