Made determining connection type more lenient
Changed the way the system checks for which type of connection type to use for the line (left/right or top/bottom). It is now more lenient and has a deadzone which allows for the current connection type to extend a bit into where the area for the opposing type was previously.