Added helper-document for pointer-tool
Added a helper for the pointer-tool (specifically for mobile touch-interaction). Implemented in a similar way to the box-selection which is fully functional. Hopefully doing this will allow for an easier conversion between mouse- and touch-inputs as they are not really communicating as is at the moment (selection and drag/drop of an element is possible with a mouse, but impossible with touch). Plan to use this is the conversion from synthetic touch to mouse-events to keep functionality with a mouse, but make it possible to do the same things on a touch-screen.