Updated setupTouchAsMouseSuppport-function
Updated the setupTouchAsMouseSupport()-function again as the previous commit didn't allow for selection and movement. The updated version makes sure to recognize selection of the parent-element even if the something inside the element has been touched. This in combination with using the "updateSelection(elData)" makes movement and seletion possible. One issue is that the element goes "invisible" when dragging it, but at this point that is a minor issue to be solved later.