fixed issue where resizing stopped working
Resizing stopped working for some reason (unsure if this was a clientside problem since it just started happening without changing any code), but the pointerstate got overridden to clicked element instead of node.

Added a condition where it couldn't happen if the pointerstate was clicked node. (Also got rid of some "tests")